+++ frontend-web/src/router/index.ts
@@ -1,5 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '@/views/HomeView.vue'
+import LearnView from '@/views/LearnView.vue'
 import PluginLabHost from '@/views/PluginLabHost.vue'
 import { getPluginRegistry } from '@/plugins/loader'
 
@@ -7,6 +8,7 @@ const plugins = getPluginRegistry()
 
 const routes = [
   { path: '/', name: 'home', component: HomeView },
+  { path: '/learn', name: 'learn', component: LearnView },
   ...plugins.map((p) => ({
     path: `${p.routePrefix}/:pathMatch(.*)*`,
     name: p.id,